% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{captionLBSPR.plots}
\alias{captionLBSPR.plots}
\title{caption LBSPR plots}
\usage{
captionLBSPR.plots(lbspr_dat, input, format = "withFig", type)
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}

\item{format}{format}

\item{type}{type}
}
\value{
something
}
\description{
caption LBSPR plots
}
\details{
something
}
