% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{captionLBSPR.tables}
\alias{captionLBSPR.tables}
\title{caption LBSPR tables}
\usage{
captionLBSPR.tables(lbspr_dat, input, format = "datatable", type)
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}

\item{format}{format}

\item{type}{type}
}
\value{
something
}
\description{
caption LBSPR tables
}
\details{
something
}
