% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dev.on}
\alias{dev.on}
\title{Device on}
\usage{
dev.on(path, format, width, height)
}
\arguments{
\item{path}{path}

\item{format}{format}

\item{width}{width}

\item{height}{height}
}
\value{
something
}
\description{
Device on
}
\details{
something
}
