% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLBSPR.data}
\alias{plotLBSPR.data}
\title{plot LBSPR data}
\usage{
plotLBSPR.data(lbspr_dat, input)
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}
}
\value{
something
}
\description{
plot LBSPR data
}
\details{
something
}
