% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLBSPR.hist}
\alias{plotLBSPR.hist}
\title{plot LBSPR hist}
\usage{
plotLBSPR.hist(lbspr_dat, input)
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}
}
\value{
something
}
\description{
plot LBSPR hist
}
\details{
something
}
