% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLBSPR.pie}
\alias{plotLBSPR.pie}
\title{plot LBSPR pie}
\usage{
plotLBSPR.pie(lbspr_dat, input)
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}
}
\value{
something
}
\description{
plot LBSPR pie
}
\details{
something
}
