% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLBSPR.sel}
\alias{plotLBSPR.sel}
\title{plot LBSPR sel}
\usage{
plotLBSPR.sel(lbspr_dat, input)
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}
}
\value{
something
}
\description{
plot LBSPR sel
}
\details{
something
}
