% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLBSPR.ts}
\alias{plotLBSPR.ts}
\title{plot LBSPR ts}
\usage{
plotLBSPR.ts(lbspr_dat, input)
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}
}
\value{
something
}
\description{
plot LBSPR ts
}
\details{
something
}
