% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lbspr.R
\name{run_lbspr}
\alias{run_lbspr}
\title{Run lbspr}
\usage{
run_lbspr(data, bin.size, linf, lm50, lm95, mk, lwa, lwb, lunit)
}
\arguments{
\item{data}{data}

\item{bin.size}{bin.size}

\item{linf}{linf}

\item{lm50}{lm50}

\item{lm95}{lm95}

\item{mk}{mk}

\item{lwa}{lwa}

\item{lwb}{lwb}

\item{lunit}{lunit}
}
\value{
something
}
\description{
Run lbspr
}
\details{
something
}
