% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{save.table}
\alias{save.table}
\title{Save table}
\usage{
save.table(file, path, format)
}
\arguments{
\item{file}{file}

\item{path}{path}

\item{format}{format}
}
\value{
something
}
\description{
Save table
}
\details{
something
}
