% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableLBSPR.data}
\alias{tableLBSPR.data}
\title{table LBSPR data}
\usage{
tableLBSPR.data(lbspr_dat, input, format = "dataframe")
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}

\item{format}{format}
}
\value{
something
}
\description{
table LBSPR data
}
\details{
something
}
