% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableLBSPR.inputPars}
\alias{tableLBSPR.inputPars}
\title{table LBSPR inputPars}
\usage{
tableLBSPR.inputPars(lbspr_dat, input, format = "datatable")
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}

\item{format}{format}
}
\value{
something
}
\description{
table LBSPR inputPars
}
\details{
something
}
