% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableLBSPR.results}
\alias{tableLBSPR.results}
\title{table LBSPR results}
\usage{
tableLBSPR.results(lbspr_dat, input, format = "datatable")
}
\arguments{
\item{lbspr_dat}{data}

\item{input}{input}

\item{format}{format}
}
\value{
something
}
\description{
table LBSPR results
}
\details{
something
}
